requirements.txt
apache-airflow==2.5.3
selenium==4.10.0
beautifulsoup4==4.12.2
pandas==2.0.3